+++ src/test/kotlin/org/wfanet/panelmatch/client/batchlookup/BatchLookupWorkflowTest.kt
@@ -51,10 +51,10 @@ class BatchLookupWorkflowTest : BeamTestBase() {
     //  - Shard 1 to have buckets 0 with "hij", 1 with "abc", and 2 with "klm"
     val parameters = Parameters(numShards = 2, numBucketsPerShard = 5, subshardSizeBytes = 1000)
 
-    val queryBundles = listOf(queryBundleOf(shard = 1, listOf(100 to 0, 101 to 0, 102 to 1)))
+    val queryBundles = listOf(queryBundleOf(shard = 1, listOf(100L to 0, 101L to 0, 102L to 1)))
 
     assertThat(runWorkflow(queryBundles, parameters))
-      .containsInAnyOrder(resultOf(100, "hij"), resultOf(101, "hij"), resultOf(102, "abc"))
+      .containsInAnyOrder(resultOf(100L, "hij"), resultOf(101L, "hij"), resultOf(102L, "abc"))
 
     assertThat(runPipelineAndGetActualMaxSubshardSize()).isEqualTo(9)
   }
@@ -68,18 +68,18 @@ class BatchLookupWorkflowTest : BeamTestBase() {
 
     val queryBundles =
       listOf(
-        queryBundleOf(shard = 0, listOf(100 to 4)),
-        queryBundleOf(shard = 1, listOf(101 to 0, 102 to 0, 103 to 1)),
-        queryBundleOf(shard = 2, listOf(104 to 0)),
+        queryBundleOf(shard = 0, listOf(100L to 4)),
+        queryBundleOf(shard = 1, listOf(101L to 0, 102L to 0, 103L to 1)),
+        queryBundleOf(shard = 2, listOf(104L to 0)),
       )
 
     assertThat(runWorkflow(queryBundles, parameters))
       .containsInAnyOrder(
-        resultOf(100, "def"),
-        resultOf(101, "hij"),
-        resultOf(102, "hij"),
-        resultOf(103, "abc"),
-        resultOf(104, "")
+        resultOf(100L, "def"),
+        resultOf(101L, "hij"),
+        resultOf(102L, "hij"),
+        resultOf(103L, "abc"),
+        resultOf(104L, "")
       )
 
     assertThat(runPipelineAndGetActualMaxSubshardSize()).isEqualTo(9)
@@ -94,18 +94,18 @@ class BatchLookupWorkflowTest : BeamTestBase() {
 
     val queryBundles =
       listOf(
-        queryBundleOf(shard = 1, listOf(100 to 0, 101 to 1, 102 to 2)),
-        queryBundleOf(shard = 1, listOf(103 to 0)),
-        queryBundleOf(shard = 1, listOf(104 to 1)),
+        queryBundleOf(shard = 1, listOf(100L to 0, 101L to 1, 102L to 2)),
+        queryBundleOf(shard = 1, listOf(103L to 0)),
+        queryBundleOf(shard = 1, listOf(104L to 1)),
       )
 
     assertThat(runWorkflow(queryBundles, parameters))
       .containsInAnyOrder(
-        resultOf(100, "hij"),
-        resultOf(101, "abc"),
-        resultOf(102, "klm"),
-        resultOf(103, "hij"),
-        resultOf(104, "abc")
+        resultOf(100L, "hij"),
+        resultOf(101L, "abc"),
+        resultOf(102L, "klm"),
+        resultOf(103L, "hij"),
+        resultOf(104L, "abc")
       )
 
     assertThat(runPipelineAndGetActualMaxSubshardSize()).isEqualTo(9)
@@ -117,8 +117,8 @@ class BatchLookupWorkflowTest : BeamTestBase() {
 
     val queryBundles =
       listOf(
-        queryBundleOf(shard = 0, listOf(1 to 53, 2 to 58)),
-        queryBundleOf(shard = 0, listOf(3 to 71, 4 to 85))
+        queryBundleOf(shard = 0, listOf(1L to 53, 2L to 58)),
+        queryBundleOf(shard = 0, listOf(3L to 71, 4L to 85))
       )
 
     runWorkflow(queryBundles, parameters)
@@ -135,16 +135,16 @@ class BatchLookupWorkflowTest : BeamTestBase() {
 
     val queryBundles =
       listOf(
-        queryBundleOf(shard = 0, listOf(1 to 53, 2 to 58)),
-        queryBundleOf(shard = 0, listOf(3 to 71, 4 to 85))
+        queryBundleOf(shard = 0, listOf(1L to 53, 2L to 58)),
+        queryBundleOf(shard = 0, listOf(3L to 71, 4L to 85))
       )
 
     assertThat(runWorkflow(queryBundles, parameters))
       .containsInAnyOrder(
-        resultOf(1, "abc"),
-        resultOf(2, "def"),
-        resultOf(3, "hij"),
-        resultOf(4, "klm")
+        resultOf(1L, "abc"),
+        resultOf(2L, "def"),
+        resultOf(3L, "hij"),
+        resultOf(4L, "klm")
       )
 
     assertThat(runPipelineAndGetActualMaxSubshardSize()).isEqualTo(3)
@@ -156,16 +156,16 @@ class BatchLookupWorkflowTest : BeamTestBase() {
 
     val queryBundles =
       listOf(
-        queryBundleOf(shard = 0, listOf(1 to 53, 2 to 58)),
-        queryBundleOf(shard = 0, listOf(3 to 71, 4 to 85))
+        queryBundleOf(shard = 0, listOf(1L to 53, 2L to 58)),
+        queryBundleOf(shard = 0, listOf(3L to 71, 4L to 85))
       )
 
     assertThat(runWorkflow(queryBundles, parameters))
       .containsInAnyOrder(
-        resultOf(1, "abc"),
-        resultOf(2, "def"),
-        resultOf(3, "hij"),
-        resultOf(4, "klm")
+        resultOf(1L, "abc"),
+        resultOf(2L, "def"),
+        resultOf(3L, "hij"),
+        resultOf(4L, "klm")
       )
 
     assertThat(runPipelineAndGetActualMaxSubshardSize()).isEqualTo(6)
@@ -175,7 +175,7 @@ class BatchLookupWorkflowTest : BeamTestBase() {
   fun `repeated bucket`() {
     val parameters = Parameters(numShards = 1, numBucketsPerShard = 1, subshardSizeBytes = 1000)
 
-    val queryBundles = listOf(queryBundleOf(shard = 0, listOf(17 to 0)))
+    val queryBundles = listOf(queryBundleOf(shard = 0, listOf(17L to 0)))
 
     assertThat(runWorkflow(queryBundles, parameters)).satisfies {
       val list = it.toList()
@@ -213,11 +213,11 @@ class BatchLookupWorkflowTest : BeamTestBase() {
   }
 }
 
-private fun resultOf(query: Int, rawPayload: String): Result {
+private fun resultOf(query: Long, rawPayload: String): Result {
   return PlaintextQueryEvaluatorTestHelper.makeResult(queryIdOf(query), rawPayload.toByteString())
 }
 
-private fun queryBundleOf(shard: Int, queries: List<Pair<Int, Int>>): QueryBundle {
+private fun queryBundleOf(shard: Int, queries: List<Pair<Long, Int>>): QueryBundle {
   return PlaintextQueryEvaluatorTestHelper.makeQueryBundle(
     shardIdOf(shard),
     queries.map { queryIdOf(it.first) to bucketIdOf(it.second) }
